Make tx buffer act like a socket
Now `lmqtt_tx_buffer_t` has a `closed` flag, which is set after it
encodes a DISCONNECT packet. If `closed` is unset and zero bytes are
processed it returns "again"; if `closed` is set and zero bytes are
processed it returns "success", which means "eof".

That makes it act like a socket and solves the problem we had in
`client_process_output` without any workaround. Moreover, the user may
now detect whether the disconnection was initiated by the client or
the server.