Write variable length buffers while decoding
Decoder used a simplified approach where a single byte was processed
per call, which proved to be very inefficient when receiving large
payloads.

Now the decoder is able to pass variable length buffers to the
underlying write method (`memcpy` or `write` depending on the
destination). Performance will be thus dependent on supplied buffer
size, instead of internal implementation.