Fix resetting of timers after end-of-file
It should be possible to reset the timers after an end-of-file is
encountered because of the following use case:

* remote closes the connection

* user waits for a period longer than the configured timeout

* user opens the socket again

* user resets the connection (lmqtt_client_reset,
  lmqtt_client_connect)

* user calls `lmqtt_client_run_once`

`lmqtt_client_reset` ignored attempts to reset the connection unless
an error condition existed (the original motivation for the method was
to clear the error state). As a result, if enough time had passed, the
client would immediately fail with a timeout error after reopening the
socket.